Create a dedicated thread for ExoPlayer
This creates ExoPlayer on its own background thread. The main issue
is that ExoPlayer requires that something (anything) on the current
thread has called Looper.prepare() before the player is created. The
two options are to blindly call Looper.prepare() on the calling thread
before creating the player, or to create a dedicated background thread
that can guarantee that Looper.prepare() has been called. Because
we don't know what side effects the former option might have, we use
the latter instead.