Make the spine element common to all player events
Previously, it was necessary to match on values of player events in
order to get at the spine element. Given that all player events will
always refer to a spine element, it makes more sense to expose that
spine element as a shared abstract property. This means that code
no longer has to pointlessly match on event values just to get the
spine element that's guaranteed to be there.