Avoid blocking on playback rate setting
If the player does not have a media codec track loaded, then attempting
to set the playback rate on the player will actually block until
a track is loaded. This adds a check to avoid actually setting the
rate unless a track is loaded. The player will pick up the correct
rate when it actually begins playing.