Improve sleep timer thread logic
This adjusts the thread logic to eliminate an exceedingly unlikely
race condition: If the timer thread task is restarted due to an
exception, and the task is resubmitted to the executor BUT the timer
is closed (and the executor shut down) before the resubmitted task runs,
the cleanup code will never run and the event status observable will
never be completed.

To fix this, the timer thread task is never resubmitted; it simply runs
forever and exceptions are handled. The cleanup code is guaranteed to
be executed.

In addition, uses latches in the test suite to hopefully make the tests
less timing sensitive (and correct the recent Travis CI test failures).

Affects: https://jira.nypl.org/browse/SIMPLY-1137