Forcibly disable RecyclerView animations
The RecyclerView has an awful misfeature: When an adapter is notified
that an item has changed, the RecyclerView will "animate" the cell that
changed as a result. It doesn't matter if anything in the cell actually
changes or not, or even if there are any visible views in the cell -
the cell will be animated. Consequently, the animated cell will bounce
around as the list is scrolled. Given that the cells in the actual
application are not being updated at a high rate (once per second,
on average), it's unclear how this animation "feature" could actually
be useful in any way for any application. I have tested the code on
three drastically different devices and the animation is broken on
all of them.

The fix is a one-line change that forcibly disables cell animation.

Fix: https://jira.nypl.org/browse/SIMPLY-1152