Fix on-device tests
The org.nypl.audiobook.android.tests.device module was
pulling in logback-classic as a transitive dependency of
org.nypl.audiobook.android.tests. This was causing the
test orchestrator on the device to fail to process the test
bytecode because of references to missing classes such as
javax.servlet.http.HttpServlet. With the proper exclusion in place,
device tests now work properly. A couple of small updates were required
to make the test suite pass.

Fix: https://jira.nypl.org/browse/SIMPLY-1102