Introduce Espresso testing
This implements the basic structure of tests using Espresso, and
includes a simple test just to verify that the configuration works.
It introduces a mock version of the audio engine API to make it
easier to test the views code in isolation.

Note that this requires devices to have their animations disabled
when testing, and this commit makes the necessary changes in Travis
CI to configure emulators as such.