Add error events
This adds a new "error" player event type, and updates the open access
player to raise them upon any detectable error. It also updates the
views to print error messages, and updates the sandbox player to
include a hidden menu that can cause the mock player implementation
to raise errors.