Adjust "skip backwards" behaviour
This changes the "skip backwards" behaviour to be consistent with
the current iOS implementation. Essentially, if the current position
is less than or equal to four seconds from the start of the current
chapter, we jump back to 15 seconds from the end of the previous
chapter (assuming one is available). If the current position is greater
than four, but less than 15 seconds from the start of the chapter,
we jump back to 00:00 in the current chapter.

Testing: All unit tests pass (some tests updated).
         Informal testing with sandbox application.

Affects: https://jira.nypl.org/browse/SIMPLY-1302