Overhaul TOC accessibility
This improves the accessibility of the TOC. All buttons are correctly
labelled. Each list item now reads the chapter number, the chapter title,
whether or not the chapter has failed to download, and whether or not the
each chapter is the currently playing chapter.

Fix: https://jira.nypl.org/browse/SIMPLY-1436