Take color schemes from the current theme
This change deprecates the ability to pass in a "primary color" to
the player fragments and instead tries to resolve a primary color
from the current application theme.

The version number has been incremented to 2.0.0 because this change
required leaning more heavily on the support library and we're unsure
as to the consequences for API compatibility that this entails.