Add a workaround for non-compat themes
The current SimplyE code cannot yet move to AppCompatActivity and
is still using the old FragmentActivity class. This, unfortunately,
means that it produces null actionViews on menu items when the action
bar is created, and this causes the current views code to crash. This
change applies a workaround; action views are inflated manually if
they are null. This has the annoying side effect of moving the action
views into a drop-down menu instead of displaying them as icons in the
action bar, but there appears to be no way to fix this. Additionally,
code in the LFA branch of SimplyE *is* using AppCompatActivity and
so this is a temporary problem; the right behaviour will occur when
the LFA branch becomes the new SimplyE code.

Affects: https://jira.nypl.org/browse/SIMPLY-1738