Send bearer tokens appropriately
This modifies the SPI implemented by extensions to allow for the
straightforward overriding of downloader provider calls, rather
than returning values which are then passed to the current download
provider. With this new mechanism in place, the Feedbooks extension
now sends the appropriate bearer tokens if a link with the correct
scheme is encountered, and configuration values have been set up for
the extension.

Affects: https://jira.nypl.org/browse/SIMPLY-2508