Feedbooks signature information is in "metadata"
Feedbooks signature and rights checking information is in the
"metadata" object instead of in the top-level manifest object.
This commit extends the extension API to allow for extensions
to register fields in either object, and updates the existing
Feedbooks manifests to use the correct arrangement.

Affects: https://jira.nypl.org/browse/SIMPLY-2507