Require a UserAgent parameter everywhere
This adds a mandatory UserAgent value to the API and uses it everywhere
an HTTP request is made. Many SPI and API methods have been changed
to take a single value of a record type rather than taking multiple
parameters. This should make it easier to extend the APIs in future
whilst preserving binary compatibility.