Stop requiring durations in manifests
This makes the changes necessary to avoid requiring manifests to
have duration information. The duration is used, if present, but the
duration reported by the player (if any) is used in preference to
this duration value.

This is needed for Overdrive audio books, which typically do not
provide duration information.